backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agenthub-backend"
version = "0.1.0"
description = "Skills Registry Platform — FastAPI Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "bcrypt>=4.1",
    "pyjwt>=2.8",
    "markdown>=3.5",
    "pyyaml>=6.0",
    "agenthub-shared",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "httpx"]

[tool.hatch.build.targets.wheel]
packages = ["app"]
